+++ src/ingest/events/post.js
@@ -6,12 +6,11 @@
 */
 
 // load node utils
-const moment = require('moment')
-const { createHashedId } = require('@swrlab/utils/packages/ard')
+const { DateTime } = require('luxon')
 
 // load eventhub utils
-const core = require('../../utils/core')
 const datastore = require('../../utils/datastore')
+const { createNewTopic, processServices } = require('../../utils/events')
 const logger = require('../../utils/logger')
 const pubsub = require('../../utils/pubsub')
 const response = require('../../utils/response')
@@ -20,73 +19,6 @@ const response = require('../../utils/response')
 const config = require('../../../config')
 
 const source = 'ingest/events/post'
-const urnPublisherRegex = /(?=urn:ard:publisher:[a-z0-9]{16})/g
-
-const processServices = async (service, req) => {
-	// fetch prefix from configured list
-	const urnPrefix = config.coreIdPrefixes[service.type]
-	const topicId = `${urnPrefix}${createHashedId(service.externalId)}`
-
-	// create hash based on prefix and id
-	service.topic = {
-		id: topicId,
-
-		// create pub/sub-compliant name
-		name: pubsub.buildId(service.topic.id),
-	}
-
-	// convert publisher if not in new urn format
-	if (!service.publisherId.match(urnPublisherRegex)) {
-		// fetch prefix
-		const urnPublisherPrefix = config.coreIdPrefixes.Publisher
-
-		// add trailing 0 if number is only 5 digits
-		if (service.publisherId.length === 5) service.publisherId = `${service.publisherId}0`
-
-		// create hash using given publisherId
-		service.publisherId = `${urnPublisherPrefix}${createHashedId(service.publisherId)}`
-	}
-
-	// fetch publisher
-	const publisher = await core.getPublisher(service.publisherId)
-
-	// block access if publisher not found
-	if (!publisher) {
-		// set blocked flag to be filtered out
-		service.blocked = `Publisher not found > ${service.publisherId}`
-
-		// log access attempt
-		logger.log({
-			level: 'warning',
-			message: service.blocked,
-			source,
-			data: { service, user: req.user },
-		})
-
-		// stop processing
-		return service
-	}
-
-	// check allowed institutions for current user
-	if (req.user.institutionId !== publisher?.institution?.id) {
-		// set blocked flag to be filtered out
-		service.blocked = 'User unauthorized for service'
-
-		// log access attempt
-		logger.log({
-			level: 'warning',
-			message: service.blocked,
-			source,
-			data: { service, user: req.user, publisher: publisher?.institution },
-		})
-
-		// stop processing
-		return service
-	}
-
-	// final data
-	return service
-}
 
 module.exports = async (req, res) => {
 	try {
@@ -99,7 +31,7 @@ module.exports = async (req, res) => {
 		}
 
 		// check offset for start event
-		if (moment(req.body.start).add(2, 'm').isBefore()) {
+		if (DateTime.now() > DateTime.fromISO(req.body.start).plus({ minutes: 2 })) {
 			return response.errors.expiredStartTime(req, res)
 		}
 
@@ -109,7 +41,7 @@ module.exports = async (req, res) => {
 		const message = {
 			name: eventName,
 			creator: req.user.email,
-			created: moment().toISOString(),
+			created: DateTime.now().toUTC().toISO(),
 
 			// use entire POST body to include potentially new fields
 			...req.body,
@@ -121,7 +53,6 @@ module.exports = async (req, res) => {
 		// save message to datastore
 		const savedMessage = await datastore.save(message, 'events')
 		message.id = savedMessage.id.toString()
-		delete message.creator
 
 		// collect unknown topics from returning errors
 		const newServices = []
@@ -137,61 +68,8 @@ module.exports = async (req, res) => {
 					service.topic.status = 'TOPIC_ERROR'
 					service.topic.messageId = null
 				} else if (messageId === 'TOPIC_NOT_FOUND') {
-					// fetch publisher
-					const publisher = await core.getPublisher(service.publisherId)
-
-					// try creating new topic
-					const newTopic = {
-						created: moment().toISOString(),
-						creator: req.user.email,
-
-						coreId: service.topic.id,
-						externalId: service.externalId,
-						name: service.topic.name,
-
-						institution: {
-							id: req.user.institutionId,
-							title: publisher.institution.title,
-						},
-
-						publisher: {
-							id: service.publisherId,
-							title: publisher.title,
-						},
-					}
-
-					// save topic to datastore
-					await datastore.save(newTopic, 'topics')
-					newTopic.id = newTopic.id.toString()
-
-					// create topic
-					const [result] = await pubsub.createTopic(newTopic)
-
-					// handle feedback
-					if (result?.name?.indexOf(service.topic.name) !== -1) {
-						// update api result that topic was created
-						service.topic.status = 'TOPIC_CREATED'
-
-						logger.log({
-							level: 'notice',
-							message: `topic created > ${service.topic.name}`,
-							source,
-							data: { service, result },
-						})
-					} else {
-						// update api result that topic was not created
-						service.topic.status = 'TOPIC_NOT_CREATED'
-
-						logger.log({
-							level: 'error',
-							message: `failed creating topic > ${service.topic.name}`,
-							source,
-							data: { service, result },
-						})
-					}
-
-					// insert empty id
-					service.topic.messageId = null
+					// first message, create a new topic
+					service.topic = await createNewTopic(service, req)
 				} else {
 					// insert messageId
 					service.topic.status = 'MESSAGE_SENT'